added glyphs
Interesting first attempt, drawing out the letterforms and adding them
to robofont. The initial challenge seems to be in the way illustrator
creates brush strokes. There are way too many extra anchor points. Not
sure how time-consuming it will be to clean them up, or if another
approach is needed in general.